BACKUP/app-api/requirements.txt
Flask
bcrypt
gunicorn
flask-cors

PyJWT
